集群不能随意删除集群

+++ spark-yun-backend/spark-yun-modules/src/main/java/com/isxcode/star/modules/cluster/service/biz/ClusterBizService.java
@@ -6,6 +6,7 @@ import com.isxcode.star.api.cluster.pojos.dto.ScpFileEngineNodeDto;
 import com.isxcode.star.api.cluster.pojos.req.*;
 import com.isxcode.star.api.cluster.pojos.res.PageClusterRes;
 import com.isxcode.star.api.cluster.pojos.res.QueryAllClusterRes;
+import com.isxcode.star.api.instance.constants.InstanceStatus;
 import com.isxcode.star.backend.api.base.exceptions.IsxAppException;
 import com.isxcode.star.common.utils.AesUtils;
 import com.isxcode.star.modules.cluster.entity.ClusterEntity;
@@ -89,6 +90,13 @@ public class ClusterBizService {
     }
 
     public void deleteCluster(DeleteClusterReq deleteClusterReq) {
+
+        // 所有的节点都卸载了，才能删除集群
+        List<ClusterNodeEntity> allNode = clusterNodeRepository.findAllByClusterId(deleteClusterReq.getEngineId());
+        boolean canNoteDelete = allNode.stream().anyMatch(e -> ClusterNodeStatus.RUNNING.equals(e.getStatus()));
+        if (canNoteDelete) {
+            throw new IsxAppException("存在节点未卸载");
+        }
 
         clusterRepository.deleteById(deleteClusterReq.getEngineId());
     }
